Replace if with single square bracket expression
The double square bracket expression has returned the result of the
comparison as exit status, i.e. on success (`$result` is zero) the
non-zero status code of the failed conditional check was returned.
A workaround would be to add a NOP at the end of the file but in this
case a single bracket expression is still clearer.

See https://gist.github.com/klingtnet/5c423f624c394f883df86bd33fd60aac
for details.