s/nobody/www-data
Unfortunately, the service won't start as `nobody`.
Both do not have shell access `/usr/sbin/nologin` but `www-data` has a
home directory, maybe this is the reason?
There was no error output, only a non-zero return code the service,
i.e. it is still to be investigated what the actual reason for the
problem is.