Rename package to `f3-server` to resolve name conflict
Unfortunately, there is already an package called `f3` in the ubuntu
repositories: https://launchpad.net/ubuntu/+source/f3

A system upgrade probably caused this `f3` to be overwritten with the
`f3` from the repositories which has a higher version number.
This resulted in a loss of the systemd service file of f3 but luckily
the service was not restarted.