Simulate directory changes
This simply stores the last path of a CD call, which always seems to be
absolute, and prepends it to the S3 objects' key.

Unfortunately, I can't test the directory change because the driver
methods are defined on a value receiver.
See [1] for details between value- and pointer-receiver structs in Go.
The implementation seems to work nonetheless.

[1]: https://nathanleclaire.com/blog/2014/08/09/dont-get-bitten-by-pointer-vs-non-pointer-method-receivers-in-golang/